requirements.txt

iniconfig==1.0.1
PyYAML==3.13
six==1.11.0
imapclient

